Refactor internal OCI package
